dump container logs to files during conformance-tests
