+++ cmd/conformance-tests/runner.go
@@ -17,20 +17,15 @@ limitations under the License.
 package main
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"encoding/xml"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"net/http"
 	"os"
-	"os/exec"
 	"path"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/go-openapi/runtime"
@@ -38,7 +33,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 
-	clusterv1alpha1 "github.com/kubermatic/machine-controller/pkg/apis/cluster/v1alpha1"
 	kubermaticapiv1 "k8c.io/kubermatic/v2/pkg/api/v1"
 	clusterclient "k8c.io/kubermatic/v2/pkg/cluster/client"
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/crd/kubermatic/v1"
@@ -53,10 +47,8 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
-	utilerror "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/apimachinery/pkg/util/rand"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/kubernetes"
@@ -64,22 +56,6 @@ import (
 	"k8s.io/client-go/util/retry"
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-func podIsReady(p *corev1.Pod) bool {
-	for _, c := range p.Status.Conditions {
-		if c.Type == corev1.PodReady {
-			return c.Status == corev1.ConditionTrue
-		}
-	}
-	return false
-}
-
-type testScenario interface {
-	Name() string
-	Cluster(secrets secrets) *apimodels.CreateClusterSpec
-	NodeDeployments(ctx context.Context, num int, secrets secrets) ([]apimodels.NodeDeployment, error)
-	OS() apimodels.OperatingSystemSpec
-}
 
 func newRunner(scenarios []testScenario, opts *Opts, log *zap.SugaredLogger) *testRunner {
 	return &testRunner{
@@ -112,23 +88,25 @@ func newRunner(scenarios []testScenario, opts *Opts, log *zap.SugaredLogger) *te
 		kubermaticProjectID:          opts.kubermaticProjectID,
 		kubermaticClient:             opts.kubermaticClient,
 		kubermaticAuthenticator:      opts.kubermaticAuthenticator,
+		containerLogsDirectory:       opts.containerLogsDirectory,
 	}
 }
 
 type testRunner struct {
-	log                *zap.SugaredLogger
-	scenarios          []testScenario
-	secrets            secrets
-	namePrefix         string
-	repoRoot           string
-	reportsRoot        string
-	PublicKeys         [][]byte
-	workerName         string
-	homeDir            string
-	printGinkoLogs     bool
-	printContainerLogs bool
-	onlyTestCreation   bool
-	pspEnabled         bool
+	log                    *zap.SugaredLogger
+	scenarios              []testScenario
+	secrets                secrets
+	namePrefix             string
+	repoRoot               string
+	reportsRoot            string
+	PublicKeys             [][]byte
+	workerName             string
+	homeDir                string
+	printGinkoLogs         bool
+	printContainerLogs     bool
+	containerLogsDirectory string
+	onlyTestCreation       bool
+	pspEnabled             bool
 
 	controlPlaneReadyWaitTimeout time.Duration
 	nodeReadyTimeout             time.Duration
@@ -367,10 +345,19 @@ func (r *testRunner) executeTests(
 	// We must store the name here because the cluster object may be nil on error
 	clusterName := cluster.Name
 
+	var logExport logExporter
+	eventExport := &eventLogger{}
+
 	if r.printContainerLogs {
+		logExport = &logPrinter{}
+	} else if r.containerLogsDirectory != "" {
+		logExport = &logDumper{directory: r.containerLogsDirectory}
+	}
+
+	if logExport != nil {
 		// Print all controlplane logs to both make debugging easier and show issues
 		// that didn't result in test failures.
-		defer r.printAllControlPlaneLogs(ctx, log, clusterName)
+		defer exportControlPlane(ctx, log, r.seedClusterClient, r.seedGeneratedClient, clusterName, logExport, eventExport)
 	}
 
 	var err error
@@ -447,8 +434,8 @@ func (r *testRunner) executeTests(
 	}
 
 	if r.printContainerLogs {
-		defer logEventsForAllMachines(ctx, log, userClusterClient)
-		defer logUserClusterPodEventsAndLogs(ctx, log, r.clusterClientProvider, cluster.DeepCopy())
+		defer exportEventsForAllMachines(ctx, log, userClusterClient, eventExport)
+		defer exportUserCluster(ctx, log, r.clusterClientProvider, cluster.DeepCopy(), logExport, eventExport)
 	}
 
 	overallTimeout := r.nodeReadyTimeout
@@ -600,48 +587,6 @@ func (r *testRunner) deleteCluster(ctx context.Context, report *reporters.JUnitT
 	return nil
 }
 
-func retryNAttempts(maxAttempts int, f func(attempt int) error) error {
-	var err error
-	for attempt := 1; attempt <= maxAttempts; attempt++ {
-		err = f(attempt)
-		if err != nil {
-			continue
-		}
-		return nil
-	}
-	return fmt.Errorf("function did not succeed after %d attempts: %v", maxAttempts, err)
-}
-
-// measuredRetryNAttempts wraps retryNAttempts with code that counts
-// the executed number of attempts and the runtimes for each
-// attempt.
-func measuredRetryNAttempts(
-	runtimeMetric *prometheus.GaugeVec,
-	//nolint:interfacer
-	attemptsMetric prometheus.Gauge,
-	log *zap.SugaredLogger,
-	maxAttempts int,
-	f func(attempt int) error,
-) func() error {
-	return func() error {
-		attempts := 0
-
-		err := retryNAttempts(maxAttempts, func(attempt int) error {
-			attempts++
-			metric := runtimeMetric.With(prometheus.Labels{"attempt": strconv.Itoa(attempt)})
-
-			return measureTime(metric, log, func() error {
-				return f(attempt)
-			})
-		})
-
-		attemptsMetric.Set(float64(attempts))
-		updateMetrics(log)
-
-		return err
-	}
-}
-
 func (r *testRunner) testCluster(
 	ctx context.Context,
 	log *zap.SugaredLogger,
@@ -754,61 +699,6 @@ func (r *testRunner) testCluster(
 	}
 
 	return nil
-}
-
-// executeGinkgoRunWithRetries executes the passed GinkgoRun and retries if it failed hard(Failed to execute the Ginkgo binary for example)
-// Or if the JUnit report from Ginkgo contains failed tests.
-// Only if Ginkgo failed hard, an error will be returned. If some tests still failed after retrying the run, the report will reflect that.
-func (r *testRunner) executeGinkgoRunWithRetries(ctx context.Context, log *zap.SugaredLogger, scenario testScenario, run *ginkgoRun, client ctrlruntimeclient.Client) (ginkgoRes *ginkgoResult, err error) {
-	const maxAttempts = 3
-
-	attempts := 1
-	defer func() {
-		ginkgoAttemptsMetric.With(prometheus.Labels{
-			"scenario": scenario.Name(),
-			"run":      run.name,
-		}).Set(float64(attempts))
-		updateMetrics(log)
-	}()
-
-	for attempts = 1; attempts <= maxAttempts; attempts++ {
-		ginkgoRes, err = r.executeGinkgoRun(ctx, log, run, client)
-
-		if ginkgoRes != nil {
-			ginkgoRuntimeMetric.With(prometheus.Labels{
-				"scenario": scenario.Name(),
-				"run":      run.name,
-				"attempt":  strconv.Itoa(attempts),
-			}).Set(ginkgoRes.duration.Seconds())
-			updateMetrics(log)
-		}
-
-		if err != nil {
-			// Something critical happened and we don't have a valid result
-			log.Errorf("Failed to execute the Ginkgo run '%s': %v", run.name, err)
-			continue
-		}
-
-		if ginkgoRes.report.Errors > 0 || ginkgoRes.report.Failures > 0 {
-			msg := fmt.Sprintf("Ginkgo run '%s' had failed tests.", run.name)
-			if attempts < maxAttempts {
-				msg = fmt.Sprintf("%s. Retrying...", msg)
-			}
-			log.Info(msg)
-			if r.printGinkoLogs {
-				if err := printFileUnbuffered(ginkgoRes.logfile); err != nil {
-					log.Infof("Error printing ginkgo logfile: %v", err)
-				}
-				log.Info("Successfully printed logfile")
-			}
-			continue
-		}
-
-		// Ginkgo run successfully and no test failed
-		return ginkgoRes, err
-	}
-
-	return ginkgoRes, err
 }
 
 func (r *testRunner) createNodeDeployments(ctx context.Context, log *zap.SugaredLogger, scenario testScenario, clusterName string) error {
@@ -1105,538 +995,4 @@ func (r *testRunner) waitUntilAllPodsAreReady(ctx context.Context, log *zap.Suga
 
 	log.Debugf("All pods became ready after %.2f seconds", time.Since(started).Seconds())
 	return nil
-}
-
-type ginkgoResult struct {
-	logfile  string
-	report   *reporters.JUnitTestSuite
-	duration time.Duration
-}
-
-const (
-	argSeparator = ` \
-    `
-)
-
-type ginkgoRun struct {
-	name       string
-	cmd        *exec.Cmd
-	reportsDir string
-	timeout    time.Duration
-}
-
-func (r *testRunner) getGinkgoRuns(
-	log *zap.SugaredLogger,
-	scenario testScenario,
-	kubeconfigFilename,
-	cloudConfigFilename string,
-	cluster *kubermaticv1.Cluster,
-) ([]*ginkgoRun, error) {
-	kubeconfigFilename = path.Clean(kubeconfigFilename)
-	repoRoot := path.Clean(r.repoRoot)
-	MajorMinor := fmt.Sprintf("%d.%d", cluster.Spec.Version.Major(), cluster.Spec.Version.Minor())
-
-	nodeNumberTotal := int32(r.nodeCount)
-
-	ginkgoSkipParallel := `\[Serial\]`
-	if minor := cluster.Spec.Version.Minor(); minor >= 16 && minor <= 20 {
-		// These require the nodes NodePort to be available from the tester, which is not the case for us.
-		// TODO: Maybe add an option to allow the NodePorts in the SecurityGroup?
-		ginkgoSkipParallel = strings.Join([]string{
-			ginkgoSkipParallel,
-			"Services should be able to change the type from ExternalName to NodePort",
-			"Services should be able to create a functioning NodePort service",
-			"Services should be able to switch session affinity for NodePort service",
-			"Services should have session affinity timeout work for NodePort service",
-			"Services should have session affinity work for NodePort service",
-		}, "|")
-	}
-
-	runs := []struct {
-		name          string
-		ginkgoFocus   string
-		ginkgoSkip    string
-		parallelTests int
-		timeout       time.Duration
-	}{
-		{
-			name:          "parallel",
-			ginkgoFocus:   `\[Conformance\]`,
-			ginkgoSkip:    ginkgoSkipParallel,
-			parallelTests: int(nodeNumberTotal) * 3,
-			timeout:       60 * time.Minute,
-		},
-		{
-			name:          "serial",
-			ginkgoFocus:   `\[Serial\].*\[Conformance\]`,
-			ginkgoSkip:    `should not cause race condition when used for configmap`,
-			parallelTests: 1,
-			timeout:       60 * time.Minute,
-		},
-	}
-	versionRoot := path.Join(repoRoot, MajorMinor)
-	binRoot := path.Join(versionRoot, "/platforms/linux/amd64")
-	var ginkgoRuns []*ginkgoRun
-	for _, run := range runs {
-
-		reportsDir := path.Join("/tmp", scenario.Name(), run.name)
-		env := []string{
-			// `kubectl diff` needs to find /usr/bin/diff
-			fmt.Sprintf("PATH=%s", os.Getenv("PATH")),
-			fmt.Sprintf("HOME=%s", r.homeDir),
-			fmt.Sprintf("AWS_SSH_KEY=%s", path.Join(r.homeDir, ".ssh", "google_compute_engine")),
-			fmt.Sprintf("LOCAL_SSH_KEY=%s", path.Join(r.homeDir, ".ssh", "google_compute_engine")),
-			fmt.Sprintf("KUBE_SSH_KEY=%s", path.Join(r.homeDir, ".ssh", "google_compute_engine")),
-		}
-
-		args := []string{
-			"-progress",
-			fmt.Sprintf("-nodes=%d", run.parallelTests),
-			"-noColor=true",
-			"-flakeAttempts=2",
-			fmt.Sprintf(`-focus=%s`, run.ginkgoFocus),
-			fmt.Sprintf(`-skip=%s`, run.ginkgoSkip),
-			path.Join(binRoot, "e2e.test"),
-			"--",
-			"--disable-log-dump",
-			fmt.Sprintf("--repo-root=%s", versionRoot),
-			fmt.Sprintf("--report-dir=%s", reportsDir),
-			fmt.Sprintf("--report-prefix=%s", run.name),
-			fmt.Sprintf("--kubectl-path=%s", path.Join(binRoot, "kubectl")),
-			fmt.Sprintf("--kubeconfig=%s", kubeconfigFilename),
-			fmt.Sprintf("--num-nodes=%d", nodeNumberTotal),
-			fmt.Sprintf("--cloud-config-file=%s", cloudConfigFilename),
-		}
-
-		args = append(args, "--provider=local")
-
-		osSpec := scenario.OS()
-		switch {
-		case osSpec.Ubuntu != nil:
-			args = append(args, "--node-os-distro=ubuntu")
-			env = append(env, "KUBE_SSH_USER=ubuntu")
-		case osSpec.Centos != nil:
-			args = append(args, "--node-os-distro=centos")
-			env = append(env, "KUBE_SSH_USER=centos")
-		case osSpec.Flatcar != nil:
-			args = append(args, "--node-os-distro=flatcar")
-			env = append(env, "KUBE_SSH_USER=core")
-		}
-
-		cmd := exec.Command(path.Join(binRoot, "ginkgo"), args...)
-		cmd.Env = env
-
-		ginkgoRuns = append(ginkgoRuns, &ginkgoRun{
-			name:       run.name,
-			cmd:        cmd,
-			reportsDir: reportsDir,
-			timeout:    run.timeout,
-		})
-	}
-
-	return ginkgoRuns, nil
-}
-
-func (r *testRunner) executeGinkgoRun(ctx context.Context, parentLog *zap.SugaredLogger, run *ginkgoRun, client ctrlruntimeclient.Client) (*ginkgoResult, error) {
-	log := parentLog.With("reports-dir", run.reportsDir)
-
-	if err := r.deleteAllNonDefaultNamespaces(ctx, log, client); err != nil {
-		return nil, fmt.Errorf("failed to cleanup namespaces before the Ginkgo run: %v", err)
-	}
-
-	timedCtx, cancel := context.WithTimeout(ctx, run.timeout)
-	defer cancel()
-
-	// We're clearing up the temp dir on every run
-	if err := os.RemoveAll(run.reportsDir); err != nil {
-		log.Errorw("Failed to remove temporary reports directory", zap.Error(err))
-	}
-	if err := os.MkdirAll(run.reportsDir, os.ModePerm); err != nil {
-		return nil, fmt.Errorf("failed to create temporary reports directory: %v", err)
-	}
-
-	// Make sure we write to a file instead of a byte buffer as the logs are pretty big
-	file, err := ioutil.TempFile("/tmp", run.name+"-log")
-	if err != nil {
-		return nil, fmt.Errorf("failed to open logfile: %v", err)
-	}
-	defer file.Close()
-	log = log.With("ginkgo-log", file.Name())
-
-	writer := bufio.NewWriter(file)
-	defer writer.Flush()
-
-	started := time.Now()
-
-	// Copy the command as we cannot execute a command twice
-	cmd := exec.CommandContext(timedCtx, "")
-	cmd.Path = run.cmd.Path
-	cmd.Args = run.cmd.Args
-	cmd.Env = run.cmd.Env
-	cmd.Dir = run.cmd.Dir
-	cmd.ExtraFiles = run.cmd.ExtraFiles
-	if _, err := writer.Write([]byte(strings.Join(cmd.Args, argSeparator))); err != nil {
-		return nil, fmt.Errorf("failed to write command to log: %v", err)
-	}
-
-	log.Infof("Starting Ginkgo run '%s'...", run.name)
-
-	// Flush to disk so we can actually watch logs
-	stopCh := make(chan struct{}, 1)
-	defer close(stopCh)
-	go wait.Until(func() {
-		if err := writer.Flush(); err != nil {
-			log.Warnw("Failed to flush log writer", zap.Error(err))
-		}
-		if err := file.Sync(); err != nil {
-			log.Warnw("Failed to sync log file", zap.Error(err))
-		}
-	}, 1*time.Second, stopCh)
-
-	cmd.Stdout = writer
-	cmd.Stderr = writer
-
-	if err := cmd.Run(); err != nil {
-		// did the context's timeout kick in?
-		if ctxErr := timedCtx.Err(); ctxErr != nil {
-			return nil, ctxErr
-		}
-
-		if exitErr, ok := err.(*exec.ExitError); ok {
-			log.Debugf("Ginkgo exited with a non-zero return code %d: %v", exitErr.ExitCode(), exitErr)
-		} else {
-			return nil, fmt.Errorf("ginkgo failed to start: %T %v", err, err)
-		}
-	}
-
-	log.Debug("Ginkgo run completed, collecting reports...")
-
-	// When running ginkgo in parallel, each ginkgo worker creates a own report, thus we must combine them
-	combinedReport, err := collectReports(run.name, run.reportsDir)
-	if err != nil {
-		return nil, err
-	}
-
-	// If we have no junit files, we cannot return a valid report
-	if len(combinedReport.TestCases) == 0 {
-		return nil, errors.New("Ginkgo report is empty, it seems no tests where executed")
-	}
-
-	combinedReport.Time = time.Since(started).Seconds()
-
-	log.Infof("Ginkgo run '%s' took %s", run.name, time.Since(started))
-	return &ginkgoResult{
-		logfile:  file.Name(),
-		report:   combinedReport,
-		duration: time.Since(started),
-	}, nil
-}
-
-func supportsStorage(cluster *kubermaticv1.Cluster) bool {
-	return cluster.Spec.Cloud.Openstack != nil ||
-		cluster.Spec.Cloud.Azure != nil ||
-		cluster.Spec.Cloud.AWS != nil ||
-		cluster.Spec.Cloud.VSphere != nil ||
-		cluster.Spec.Cloud.GCP != nil ||
-		cluster.Spec.Cloud.Hetzner != nil
-}
-
-func supportsLBs(cluster *kubermaticv1.Cluster) bool {
-	return cluster.Spec.Cloud.Azure != nil ||
-		cluster.Spec.Cloud.AWS != nil ||
-		cluster.Spec.Cloud.GCP != nil ||
-		(cluster.Spec.Cloud.Hetzner != nil && cluster.Spec.Version.Minor() >= 18)
-}
-
-func (r *testRunner) printAllControlPlaneLogs(ctx context.Context, log *zap.SugaredLogger, clusterName string) {
-	log.Info("Printing control plane logs")
-
-	cluster := &kubermaticv1.Cluster{}
-	if err := r.seedClusterClient.Get(ctx, types.NamespacedName{Name: clusterName}, cluster); err != nil {
-		log.Errorw("Failed to get cluster", zap.Error(err))
-		return
-	}
-
-	log.Debugw("Cluster health status", "status", cluster.Status.ExtendedHealth)
-
-	log.Info("Logging events for cluster")
-	if err := logEventsObject(ctx, log, r.seedClusterClient, "default", cluster.UID); err != nil {
-		log.Errorw("Failed to log cluster events", zap.Error(err))
-	}
-
-	if err := printEventsAndLogsForAllPods(
-		ctx,
-		log,
-		r.seedClusterClient,
-		r.seedGeneratedClient,
-		cluster.Status.NamespaceName,
-	); err != nil {
-		log.Errorw("Failed to print events and logs of pods", zap.Error(err))
-	}
-}
-
-// waitForMachinesToJoinCluster waits for machines to join the cluster. It does so by checking
-// if the machines have a nodeRef. It does not check if the nodeRef is valid.
-// All errors are swallowed, only the timeout error is returned.
-func waitForMachinesToJoinCluster(ctx context.Context, log *zap.SugaredLogger, client ctrlruntimeclient.Client, timeout time.Duration) (time.Duration, error) {
-	startTime := time.Now()
-	err := wait.Poll(10*time.Second, timeout, func() (bool, error) {
-		machineList := &clusterv1alpha1.MachineList{}
-		if err := client.List(ctx, machineList); err != nil {
-			log.Warnw("Failed to list machines", zap.Error(err))
-			return false, nil
-		}
-		for _, machine := range machineList.Items {
-			if !machineHasNodeRef(machine) {
-				log.Infow("Machine has no nodeRef yet", "machine", machine.Name)
-				return false, nil
-			}
-		}
-		log.Infow("All machines got a Node", "duration-in-seconds", time.Since(startTime).Seconds())
-		return true, nil
-	})
-	return timeout - time.Since(startTime), err
-}
-
-func machineHasNodeRef(machine clusterv1alpha1.Machine) bool {
-	return machine.Status.NodeRef != nil && machine.Status.NodeRef.Name != ""
-}
-
-// WaitForNodesToBeReady waits for all nodes to be ready. It does so by checking the Nodes "Ready"
-// condition. It swallows all errors except for the timeout.
-func waitForNodesToBeReady(ctx context.Context, log *zap.SugaredLogger, client ctrlruntimeclient.Client, timeout time.Duration) (time.Duration, error) {
-	startTime := time.Now()
-	err := wait.Poll(10*time.Second, timeout, func() (bool, error) {
-		nodeList := &corev1.NodeList{}
-		if err := client.List(ctx, nodeList); err != nil {
-			log.Warnw("Failed to list nodes", zap.Error(err))
-			return false, nil
-		}
-		for _, node := range nodeList.Items {
-			if !nodeIsReady(node) {
-				log.Infow("Node is not ready", "node", node.Name)
-				return false, nil
-			}
-		}
-		log.Infow("All nodes got ready", "duration-in-seconds", time.Since(startTime).Seconds())
-		return true, nil
-	})
-	return timeout - time.Since(startTime), err
-}
-
-func nodeIsReady(node corev1.Node) bool {
-	for _, c := range node.Status.Conditions {
-		if c.Type == corev1.NodeReady {
-			return c.Status == corev1.ConditionTrue
-		}
-	}
-	return false
-}
-
-func printFileUnbuffered(filename string) error {
-	fd, err := os.Open(filename)
-	if err != nil {
-		return err
-	}
-	defer fd.Close()
-	return printUnbuffered(fd)
-}
-
-// printUnbuffered uses io.Copy to print data to stdout.
-// It should be used for all bigger logs, to avoid buffering
-// them in memory and getting oom killed because of that.
-func printUnbuffered(src io.Reader) error {
-	_, err := io.Copy(os.Stdout, src)
-	return err
-}
-
-// junitReporterWrapper is a convenience func to get junit results for a step
-// It will create a report, append it to the passed in testsuite and propagate
-// the error of the executor back up
-// TODO: Should we add optional retrying here to limit the amount of wrappers we need?
-func junitReporterWrapper(
-	testCaseName string,
-	report *reporters.JUnitTestSuite,
-	executor func() error,
-	extraErrOutputFn ...func() string,
-) error {
-	junitTestCase := reporters.JUnitTestCase{
-		Name:      testCaseName,
-		ClassName: testCaseName,
-	}
-
-	startTime := time.Now()
-	err := executor()
-	junitTestCase.Time = time.Since(startTime).Seconds()
-	if err != nil {
-		junitTestCase.FailureMessage = &reporters.JUnitFailureMessage{Message: err.Error()}
-		report.Failures++
-		for _, extraOut := range extraErrOutputFn {
-			extraOutString := extraOut()
-			err = fmt.Errorf("%v\n%s", err, extraOutString)
-			junitTestCase.FailureMessage.Message += "\n" + extraOutString
-		}
-	}
-
-	report.TestCases = append(report.TestCases, junitTestCase)
-	report.Tests++
-
-	return err
-}
-
-// printEvents and logs for all pods. Include ready pods, because they may still contain useful information.
-func printEventsAndLogsForAllPods(
-	ctx context.Context,
-	log *zap.SugaredLogger,
-	client ctrlruntimeclient.Client,
-	k8sclient kubernetes.Interface,
-	namespace string,
-) error {
-	log.Infow("Printing logs for all pods", "namespace", namespace)
-
-	pods := &corev1.PodList{}
-	if err := client.List(ctx, pods, ctrlruntimeclient.InNamespace(namespace)); err != nil {
-		return fmt.Errorf("failed to list pods: %v", err)
-	}
-
-	var errs []error
-	for _, pod := range pods.Items {
-		log := log.With("pod", pod.Name)
-		if !podIsReady(&pod) {
-			log.Error("Pod is not ready")
-		}
-		log.Info("Logging events for pod")
-		if err := logEventsObject(ctx, log, client, pod.Namespace, pod.UID); err != nil {
-			log.Errorw("Failed to log events for pod", zap.Error(err))
-			errs = append(errs, err)
-		}
-		log.Info("Printing logs for pod")
-		if err := printLogsForPod(ctx, log, k8sclient, &pod); err != nil {
-			log.Errorw("Failed to print logs for pod", zap.Error(utilerror.NewAggregate(err)))
-			errs = append(errs, err...)
-		}
-	}
-
-	return utilerror.NewAggregate(errs)
-}
-
-func printLogsForPod(ctx context.Context, log *zap.SugaredLogger, k8sclient kubernetes.Interface, pod *corev1.Pod) []error {
-	var errs []error
-	for _, container := range pod.Spec.Containers {
-		containerLog := log.With("container", container.Name)
-		containerLog.Info("Printing logs for container")
-		if err := printLogsForContainer(ctx, k8sclient, pod, container.Name); err != nil {
-			containerLog.Errorw("Failed to print logs for container", zap.Error(err))
-			errs = append(errs, err)
-		}
-	}
-	for _, initContainer := range pod.Spec.InitContainers {
-		containerLog := log.With("initContainer", initContainer.Name)
-		containerLog.Infow("Printing logs for initContainer")
-		if err := printLogsForContainer(ctx, k8sclient, pod, initContainer.Name); err != nil {
-			containerLog.Errorw("Failed to print logs for initContainer", zap.Error(err))
-			errs = append(errs, err)
-		}
-	}
-	return errs
-}
-
-func printLogsForContainer(ctx context.Context, client kubernetes.Interface, pod *corev1.Pod, containerName string) error {
-	readCloser, err := client.
-		CoreV1().
-		Pods(pod.Namespace).
-		GetLogs(pod.Name, &corev1.PodLogOptions{Container: containerName}).
-		Stream(ctx)
-	if err != nil {
-		return err
-	}
-	defer readCloser.Close()
-	return printUnbuffered(readCloser)
-}
-
-func logEventsForAllMachines(
-	ctx context.Context,
-	log *zap.SugaredLogger,
-	client ctrlruntimeclient.Client,
-) {
-	machines := &clusterv1alpha1.MachineList{}
-	if err := client.List(ctx, machines); err != nil {
-		log.Errorw("Failed to list machines", zap.Error(err))
-		return
-	}
-
-	for _, machine := range machines.Items {
-		machineLog := log.With("name", machine.Name)
-		machineLog.Infow("Logging events for machine")
-		if err := logEventsObject(ctx, log, client, machine.Namespace, machine.UID); err != nil {
-			machineLog.Errorw("Failed to log events for machine", "namespace", machine.Namespace, zap.Error(err))
-		}
-	}
-}
-
-func logEventsObject(
-	ctx context.Context,
-	log *zap.SugaredLogger,
-	client ctrlruntimeclient.Client,
-	namespace string,
-	uid types.UID,
-) error {
-	events := &corev1.EventList{}
-	listOpts := &ctrlruntimeclient.ListOptions{
-		Namespace:     namespace,
-		FieldSelector: fields.OneTermEqualSelector("involvedObject.uid", string(uid)),
-	}
-	if err := client.List(ctx, events, listOpts); err != nil {
-		return fmt.Errorf("failed to get events: %v", err)
-	}
-
-	for _, event := range events.Items {
-		var msg string
-		if event.Type == corev1.EventTypeWarning {
-			// Make sure this gets highlighted
-			msg = "ERROR"
-		}
-		log.Infow(
-			msg,
-			"EventType", event.Type,
-			"Number", event.Count,
-			"Reason", event.Reason,
-			"Message", event.Message,
-			"Source", event.Source.Component,
-		)
-	}
-	return nil
-}
-
-func logUserClusterPodEventsAndLogs(
-	ctx context.Context,
-	log *zap.SugaredLogger,
-	connProvider *clusterclient.Provider,
-	cluster *kubermaticv1.Cluster,
-) {
-	log.Info("Attempting to log usercluster pod events and logs")
-	cfg, err := connProvider.GetClientConfig(ctx, cluster)
-	if err != nil {
-		log.Errorw("Failed to get usercluster admin kubeconfig", zap.Error(err))
-		return
-	}
-	k8sClient, err := kubernetes.NewForConfig(cfg)
-	if err != nil {
-		log.Errorw("Failed to construct k8sClient for usercluster", zap.Error(err))
-		return
-	}
-	client, err := connProvider.GetClient(ctx, cluster)
-	if err != nil {
-		log.Errorw("Failed to construct client for usercluster", zap.Error(err))
-		return
-	}
-	if err := printEventsAndLogsForAllPods(
-		ctx,
-		log,
-		client,
-		k8sClient,
-		"",
-	); err != nil {
-		log.Errorw("Failed to print events and logs for usercluster pods", zap.Error(err))
-	}
 }
